+++ vsu-app/Startup.cs
@@ -1,21 +1,10 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-
 using Gov.Cscp.Victims.Public.Services;
-using Gov.Cscp.Victims.Public.Models;
-
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.CookiePolicy;
-using Microsoft.AspNetCore.DataProtection;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Features;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Authorization;
 
 namespace Gov.Cscp.Victims.Public
 {
@@ -31,9 +20,10 @@ namespace Gov.Cscp.Victims.Public
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            // add singleton to allow Controllers to query the Request object
-            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-            services.AddSingleton<IDynamicsResultService, DynamicsResultService>();
+            services.AddSingleton<ICOASTAuthService, COASTAuthService>();
+            services.AddTransient<TokenHandler>();
+
+            services.AddHttpClient<IDynamicsResultService, DynamicsResultService>().AddHttpMessageHandler<TokenHandler>();
 
             services.AddControllersWithViews();
             // In production, the Angular files will be served from this directory
